requirements.txt
keras==2.3.1
scikit-learn==0.22.1
scikit-image~=0.17.2
matplotlib~=3.3.4
nibabel~=3.2.1
medpy~=0.4.0
pandas~=1.1.5
seaborn~=0.11.2
numpy~=1.18.5
click~=8.0.3
tensorflow-gpu==2.3.0